remove lo_k, moving shared functions between kbe/lo_k and mbe/lo to shared/external/lo_helper
